requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
asyncio==3.4.3
attrs==23.1.0
catalogue==2.0.8
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
confection==0.0.4
cymem==2.0.7
DAWG-Python==0.7.2
distlib==0.3.6
docopt==0.6.2
docutils==0.20.1
filelock==3.12.0
frozenlist==1.3.3
fsspec==2023.5.0
idna==3.4
Jinja2==3.1.2
joblib==1.2.0
Kivy-Garden==0.1.5
langcodes==3.3.0
MarkupSafe==2.1.2
mpmath==1.3.0
multidict==6.0.4
murmurhash==1.0.9
networkx==3.1
numpy==1.24.3
packaging==23.1
pathy==0.10.1
pipenv==2023.2.18
platformdirs==3.5.1
preshed==3.0.8
psycopg2-binary==2.9.6
pydantic==1.10.7
Pygments==2.15.1
pymorphy3==1.2.0
pymorphy3-dicts-ru==2.4.417150.4580142
PyPika==0.48.9
python-dotenv==1.0.0
PyYAML==6.0
redis==4.5.5
regex==2023.5.5
requests==2.30.0
scikit-learn==1.2.2
scipy==1.10.1
sentencepiece==0.1.99
sklearn==0.0.post5
smart-open==6.3.0
spacy==3.5.3
spacy-legacy==3.0.12
spacy-loggers==1.0.4
srsly==2.4.6
sympy==1.12
thinc==8.1.10
threadpoolctl==3.1.0
tokenizers==0.13.3
torch==2.0.1
tqdm==4.65.0
transformers==4.29.2
typer==0.7.0
typing_extensions==4.5.0
urllib3==2.0.2
virtualenv==20.23.0
virtualenv-clone==0.5.7
wasabi==1.1.1
yarl==1.9.2
